pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "doc_processor"
version = "0.1.0"
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
description = "车辆数据文档处理工具"
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
dependencies = [
    "python-docx>=0.8.11",
    "click>=8.0.0",
    "rich>=10.0.0",
    "pandas>=1.3.0",
    "lxml>=4.6.3",
    "PyYAML>=6.0",
    "psutil>=5.9.0",
]

[project.urls]
"Homepage" = "https://github.com/yourusername/doc-processor"
"Bug Tracker" = "https://github.com/yourusername/doc-processor/issues"

[project.scripts]
doc-processor = "doc_processor.cli.commands:cli"

[tool.setuptools]
packages = [
    "doc_processor",
    "doc_processor.cli",
    "doc_processor.config",
    "doc_processor.core",
    "doc_processor.exceptions",
    "doc_processor.extractors",
    "doc_processor.models",
    "doc_processor.parsers",
    "doc_processor.utils"
]

[tool.black]
line-length = 88
target-version = ["py38"]

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true 
